requirements.txt
Flask
tensorflow-cpu
Pillow
gunicorn
numpy